[generator] ClusterFinder was moved to cluster_finder.hpp.
